rabbitmq/worker/requirements.txt
pika==1.1.0
requests==2.22.0
Pillow==7.0.0